Fix feature config
